Move setup() to its own file; rename to vanillaSetup
